# [Dispatcher] improve api, reduce overhead, improve performances for items > 1k