--- a/sources/core/Stride.Core/Threading/ThreadPool.SemaphoreW.cs
+++ b/sources/core/Stride.Core/Threading/ThreadPool.SemaphoreW.cs
@@ -15,7 +15,7 @@ namespace Stride.Core.Threading
         /// <summary>
         /// Mostly lifted from dotnet's LowLevelLifoSemaphore
         /// </summary>
-        private class SemaphoreW
+        private class SemaphoreW : ISemaphore
         {
             private const int SpinSleep0Threshold = 10;
             
@@ -61,6 +61,8 @@ namespace Stride.Core.Threading
             public void Wait(int timeout = -1) => internals.Wait(spinCount, lifoSemaphore, timeout);
 
             public void Release(int releaseCount) => internals.Release(releaseCount, lifoSemaphore);
+
+            public void Dispose() => lifoSemaphore?.Dispose();
 
             [StructLayout(LayoutKind.Explicit)]
             private struct Counts