… sources/engine/Stride.Rendering/Rendering/Materials/MaterialRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/SkinningRenderFeature.cs
… sources/engine/Stride.Rendering/Rendering/TransformRenderFeature.cs