@Eideren 2023-12-20
Here's a benchmark running on latest changes:

| Method  | ItemsToProcess | Mean       | Error     | StdDev     |
|-------- |--------------- |------------|-----------|------------|
| New     | 10             |   1.093 us | 0.0215 us |  0.0393 us |
| Old     | 10             |   5.357 us | 0.1058 us |  0.2038 us |
| NewLong | 10             |   1.383 us | 0.0277 us |  0.0477 us |
| OldLong | 10             |   5.525 us | 0.0521 us |  0.0435 us |
| New     | 100            |   1.453 us | 0.0207 us |  0.0162 us |
| Old     | 100            |   2.735 us | 0.0528 us |  0.0494 us |
| NewLong | 100            |   4.312 us | 0.0857 us |  0.0880 us |
| OldLong | 100            |   7.431 us | 0.1068 us |  0.0892 us |
| New     | 1000           |   2.084 us | 0.0414 us |  0.0960 us |
| Old     | 1000           |   7.699 us | 0.1528 us |  0.2510 us |
| NewLong | 1000           |   8.039 us | 0.0244 us |  0.0228 us |
| OldLong | 1000           |   9.355 us | 0.1729 us |  0.2123 us |
| New     | 10000          |   6.890 us | 0.1352 us |  0.2761 us |
| Old     | 10000          |   8.564 us | 0.0888 us |  0.0693 us |
| NewLong | 10000          |  23.778 us | 0.1935 us |  0.1715 us |
| OldLong | 10000          |  23.918 us | 0.1049 us |  0.0876 us |
| New     | 100000         |  29.414 us | 0.2818 us |  0.2636 us |
| Old     | 100000         |  33.156 us | 0.6590 us |  1.3162 us |
| NewLong | 100000         | 193.900 us | 1.9274 us |  1.5048 us |
| OldLong | 100000         | 232.131 us | 7.2847 us | 21.4791 us |

Difference between `New` and `NewLong` is that more operations are run per item for the latter. It doesn't matter too much in this case since both the old and new use the same job splitting tactic but it does give some insight into how much overhead there is to just schedule a job.
[Benchmark.zip](https://github.com/stride3d/stride/files/13732288/Benchmark.zip)